backend/requirements.txt
fastapi
uvicorn
langchain
langchain-fireworks
pymongo
pyarrow==19.0.0
pydantic
langchain-mongodb
python-dotenv
requests
azure-core
azure-devops
msrest
kink
google-cloud-firestore
langchain-google-firestore
langchain-google-vertexai
functions-framework
google-events
google-genai